LayoutTests/fast/images/image-extraction/mac/copy-image-overlay-text-with-context-menu-expected.txt

PASS editor.textContent is "foobar"
PASS successfullyParsed is true

TEST COMPLETE
